🧪 Use `reusable-tox.yml` @ `tox-dev/workflow`
This patch integrates the upstream workflow hosted at [[1]]. It
features a set of steps that are common for when having tox-centric
CI configurations. In addition to that, it also integrates automatic
verbose reruns on test failures and publishing coverage and test
results to Codecov (currently disabled).

The reusable workflow is customizable though hook points that run
in-repo actions with defined names hosted under
`.github/reusables/tox-dev/workflow/reusable-tox/hooks/`.

This change adds two hooks to embed container image build steps
specific to integration test jobs.

Among small functional changes, there's a slight optimization that
skips installing `pypa/build` and uses the readily available
`pip wheel` command. The patch also includes an update to the yamllint
config that allows any consistent indents for YAML sequences. This is
useful since the changes made here are periodically synchronized with
a number of other projects.

Additionally, an integration is added with `alls-green` [[2]] that
lets us migrate to using just a single branch protection check for the
CI instead of enumerating each possible job name in there.

The usage of `tox-uv` improved the tox env provisioning speed and
built-in caching does the same on the pip level.

[1]: https://github.com/tox-dev/workflow
[2]: https://github.com/marketplace/actions/alls-green#why