Python 3.14 compat: replace codecs.open with open (#1434)
Python 3.14 deprecates codecs.open:
https://docs.python.org/3.14/whatsnew/3.14.html#deprecated
This makes the tests fail, because we get a bunch of warnings
about the deprecation and I guess the tests are set to fail on
warnings.

It's never been recommended to use codecs.open on Python 3. It
was the "right way" to open text content on Python 2 for a bit,
but when Python 3 came around, io.open was supposed to be used
to be agnostic between 2 and 3. Since ansible-runner only targets
3 nowadays, let's just use open.

Signed-off-by: Adam Williamson <awilliam@redhat.com>